+++ modulo3/labex/src/Pages/AdminHomePage.js
@@ -1,25 +1,72 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
-import { goToBackHome, goToCreateTrip, goToLogin } from "../routes/coordinator";
-import {useProtectedPage} from "../hooks/useProtected";
-
-
+import { goToBackHome, goToCreateTrip, goToDetailsTrips } from "../routes/coordinator";
+import { useProtectedPage } from "../hooks/useProtected";
+import { Url } from "../constants/Url";
+import axios from "axios";
+import { Container, Second } from "./StyleTodos/AdminStyle"
 
 
 export const AdminHomePage = () => {
-    useProtectedPage()    
+    const [listTrips, setLisTrips] = useState([])
+    useProtectedPage()
+    const token = localStorage.getItem('token')
+
+
+    useProtectedPage()
     const navigate = useNavigate()
 
     const useClearStorage = () => {
         window.localStorage.clear('token')
+        navigate("/login")
     }
 
+    useEffect(() => {
+        axios.get(`${Url}/trips`)
+            .then((response) => {
+                setLisTrips(response.data.trips);
+            })
+            .catch((err) => {
+                console.log("Deu erro", err.response)
+            })
+    }, [listTrips])
+
+    const DelTrip = (id) => {
+        if (window.confirm("Tem certeza que deseja deletar?")) {
+            axios.delete(`${Url}/trips/${id}`, {
+                headers: {
+                    "Content-Type": "application/json",
+                    auth: token
+                }
+            })
+                .then((response) => {
+                    console.log("Excluiu");
+                })
+                .catch((err) => {
+                    console.log(err);
+                })
+        }
+
+    }
+
+    const render = listTrips.map((list) => {
+        return (
+            <div className="Lista" key={list.id}>
+                <p onClick={() => goToDetailsTrips(navigate, list.id)}>Nome: {list.name}</p>
+                <button onClick={() => DelTrip(list.id)}>X</button>
+            </div>
+        )
+    })
+
     return (
-        <div>
+        <Container>
             <h2>Painel Administrativo</h2>
-            <button onClick={() => goToBackHome(navigate)}>Voltar</button>
-            <button onClick={() => goToCreateTrip(navigate)}>Criar Viagem</button>
-            <button onClick={ useClearStorage }> Logout</button>
-        </div>
+            <Second>
+                <button onClick={() => goToBackHome(navigate)}>Voltar</button>
+                <button onClick={() => goToCreateTrip(navigate)}>Criar Viagem</button>
+                <button onClick={useClearStorage}>Logout</button>
+            </Second>
+            {render}
+        </Container>
     )
 }
