+++ modulo3/labex/src/Pages/HomePage.js
@@ -1,8 +1,6 @@
 import React from "react";
 import { useNavigate } from "react-router-dom";
-import { Url } from "../constants/Url";
-import { useEffect } from "react";
-import axios from "axios";
+import { Main, Second } from "./StyleTodos/HomeStyle"
 import { goToListTripsPage, goToAdmin } from "../routes/coordinator";
 
 
@@ -10,17 +8,19 @@ import { goToListTripsPage, goToAdmin } from "../routes/coordinator";
 export const HomePage = () => {
     const navigate = useNavigate()
 
-   
+
 
 
     return (
-        <div>
-            <h2>Home</h2>
+        <Main>
+            <Second>
+                <h2>Home</h2>
 
-            <div>
-                <button onClick={() => goToListTripsPage(navigate)}>Ver Viagens </button>
-                <button onClick={() => goToAdmin(navigate)}> Área de Admin </button>
-            </div>
-        </div>
+                <div>
+                    <button onClick={() => goToListTripsPage(navigate)}>Ver Viagens </button>
+                    <button onClick={() => goToAdmin(navigate)}> Área de Admin </button>
+                </div>
+            </Second>
+        </Main>
     )
 }
