+++ modulo3/labex/src/Pages/ListTripsPage.js
@@ -3,44 +3,45 @@ import React, { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { Url } from "../constants/Url";
 import { goToHomePage, goToApplication } from "../routes/coordinator";
-
+import {Container} from "./StyleTodos/ListTripsStyle"
 
 export const ListTripsPage = () => {
-const [listTrips, setLisTrips] = useState([])
+    const [listTrips, setLisTrips] = useState([])
 
-const navigate = useNavigate()
+    const navigate = useNavigate()
 
-useEffect(() => {
-    axios.get(`${Url}/trips`)
-    .then((response) => {
-        setLisTrips(response.data.trips);
+    useEffect(() => {
+        axios.get(`${Url}/trips`)
+            .then((response) => {
+                setLisTrips(response.data.trips);
+            })
+            .catch((err) => {
+                console.log("Deu erro", err.response)
+            })
+    }, [listTrips])
+
+    const render = listTrips.map((list) => {
+        return (
+            <div className="Lista" key={list.id}>
+                <p>Nome: {list.name} </p>
+                <p>Descrição: {list.description} </p>
+                <p>Planeta: {list.planet} </p>
+                <p>duração: {list.durationInDays} </p>
+                <p>Data: {list.date} </p>
+            </div>
+        )
     })
-    .catch((err) => {
-        console.log("Deu erro", err.response)
-    })
-}, [listTrips])
-
-const render = listTrips.map((list) => {
-    return (
-        <div className="Lista" key={list.id}>            
-            <p>Nome: {list.name} </p>
-            <p>Descrição: {list.description} </p>
-            <p>Planeta: {list.planet} </p>
-            <p>duração: {list.durationInDays} </p>
-            <p>Data: {list.date} </p>
-        </div>
-    )
-})
 
     return (
-        <div>
+        <Container>
             <h2>Lista de Viagens</h2>
 
             <div>
                 <button onClick={() => goToApplication(navigate)}>Inscrever-se</button>
-                <button onClick={() => goToHomePage(navigate)}>Voltar</button>                
-                {render}
+                <button onClick={() => goToHomePage(navigate)}>Voltar</button>
+                
             </div>
-        </div>
+            {render}
+        </Container>
     )
 }
