+++ modulo3/labex/src/Pages/StyleTodos/HomeStyle.js
@@ -1,0 +1,22 @@
+import styled from "styled-components";
+
+export const Main = styled.div`
+display: flex;
+text-align: center;
+height: 100vh;
+
+
+`
+export const Second = styled.div`
+display: flex;
+flex-direction: column;
+justify-content: center;
+width: 50%;
+height: 200px;
+margin: auto;
+box-shadow: 5px 5px 5px black;
+
+button{
+    margin: 0 10px;
+}
+`
