+++ modulo3/labex/src/routes/Router.js
@@ -20,8 +20,8 @@ export const Router = () => {
           <Route path="/trips/application" element={<ApplicationFormPage />} />
           <Route path="/login" element={<LoginPage />} />
           <Route path="/admin/trips/list" element={<AdminHomePage />} />
-          <Route path="/admin/trips/create" element={ <TripDetailsPage />} />
-          <Route path="/admin/trips/:id" element={  <CreateTripPage />} />         
+          <Route path="/admin/trips/:id" element={ <TripDetailsPage />} />
+          <Route path="/admin/trips/create" element={  <CreateTripPage />} />         
         </Routes>
       </BrowserRouter>
     )
